backend/requirements.txt
fastapi
uvicorn[standard]
surrealdb
langgraph
langchain
langchain-openai
python-dotenv

